Routes refactoring
